Bus v.0.6.1: change bus_number field data type

+++ database/migrations/2018_11_02_162843_create_bus_table.php
@@ -15,7 +15,7 @@ class CreateBusTable extends Migration
     {
         Schema::create('bus', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('bus_number');
+            $table->smallInteger('bus_number');
             $table->bigInteger('price');
             $table->integer('bus_admin_id')->unsigned();
             $table->timestamps();
